requirements.txt
gradio==3.43.2
diffusers==0.20.2
transformers==4.28.0
accelerate==0.23.0
python-dotenv==1.0.1
numpy<2.0
huggingface-hub<0.20.0
matplotlib>=3.5.0
Pillow>=8.3.0
tqdm>=4.62.0